Put video clips together with Moviepy
Noticed that the recorded duration in `.json` and the duration of the clip can vary.
This is especially bad with very short test cases (5 seconds recorded vs 1 second video file).
With longer clips the differences are less serious.

As a workaround for this each test case has a pause in the beginning and end of the test case.
The narration times seem to be mostly right, but it's worth considering a different screencast recording implementation.